* 'develop' of https://github.com/bigquann97/BusNo10-Board: (49 commits)
  [refactor] ManyToOne 객체 참조로 리팩터링
  [feat] My-Info 기능 추가
  [feat] 대댓글 기능 구현
  [feat] 대댓글 구현
  [feat] 대댓글 기능
  [fix] 주석 코드 정리
  [fix] 변수명 수정
  [fix] 좋아요API URL 재설계. Like Controller, Service 추가
  [feat] Comment에 Like 기능 구현
  [feat] Post에 Like 기능 구현
  [feat] #23 ResponseDto Id 변수 추가
  [feat] Post 삭제 코드 수정, Post api 시큐리티 권한 적용
  [fix] Post 엔티티에서 username변수가 unique로 설정되어 있던 것 제거(User 엔티티에서는 unique이지만 Post 엔티티에서는 중복이 가능해야함.)
  [fix] conflict해결
  [feat] 대댓글 기능 구현
  [feat] 대댓글 구현
  [feat] 대댓글 기능
  [fix] 주석 코드 정리
  [fix] 변수명 수정
  [fix] 좋아요API URL 재설계. Like Controller, Service 추가
  ...

+++ src/main/java/sparta/bus10/service/PostService.java
@@ -2,46 +2,79 @@ package sparta.bus10.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import sparta.bus10.dto.PostRequestDto;
 import sparta.bus10.dto.PostResponseDto;
+import sparta.bus10.entity.Comment;
+import sparta.bus10.entity.Like;
 import sparta.bus10.entity.Post;
+import sparta.bus10.entity.User;
+import sparta.bus10.repository.CommentRepository;
+import sparta.bus10.repository.LikeRepository;
 import sparta.bus10.repository.PostRepository;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class PostService {
+
     private final PostRepository postRepository;
-    public void createService(PostRequestDto requestDto) {
-        postRepository.save(new Post(requestDto));
+    private final CommentRepository commentRepository;
+
+    @Transactional
+    public void createPost(PostRequestDto postrequestDto, User user) {
+        System.out.println(user.getUsername());
+        Post post = new Post(user, postrequestDto.getPostTitle(), postrequestDto.getPostContent());
+        postRepository.save(post);
     }
 
-    public List<PostResponseDto> getPostAll(){
-        List<Post> post = postRepository.findAll();
-        return post.stream().map(x-> new PostResponseDto(x)).collect(Collectors.toList());
+    @Transactional
+    public List<PostResponseDto> getPostAll() {
+        List<Post> posts = postRepository.findAll();
+        List<PostResponseDto> resultPosts = new ArrayList<>();
+        for (Post post : posts) {
+            List<Comment> comments = commentRepository.findByPost(post);
+            PostResponseDto postResponseDto = new PostResponseDto(post, comments);
+            resultPosts.add(postResponseDto);
+        }
+        return resultPosts;
     }
 
-
-    public PostResponseDto getPostOne(Long postId){
+    @Transactional
+    public PostResponseDto getPostOne(Long postId) {
         Post post = postRepository.findById(postId).orElseThrow(
-            ()-> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
-//        PostResponseDto postResponseDto = new PostResponseDto(post);
-        return new PostResponseDto(post);
+                () -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
+        List<Comment> comment = commentRepository.findByPost(post);
+        return new PostResponseDto(post, comment);
     }
 
-    public void editPost(Long postId ,PostRequestDto postrequestDto) {
-       Post post = postRepository.findById(postId).orElseThrow(
-               ()-> new IllegalArgumentException("게시글을 찾을 수 없습니다.")
-       );
-       post.changePost(postrequestDto);
-       postRepository.save(post);
-    }
-    public void deletePost(Long postId){
+    @Transactional
+    public void editPost(Long postId, PostRequestDto postrequestDto, User user) {
         Post post = postRepository.findById(postId).orElseThrow(
-                ()-> new IllegalArgumentException("게시글을 찾을 수 없습니다.")
+                () -> new IllegalArgumentException("게시글을 찾을 수 없습니다.")
         );
+        if (!post.validateUser(user)) {
+            throw new IllegalArgumentException("유저의 이름이 일치하지 않습니다.");
+        }
+        post.changePost(user, postrequestDto.getPostTitle(), postrequestDto.getPostContent());
+        postRepository.save(post);
+    }
+
+    @Transactional
+    public void deletePost(Long postId, User user) {
+        Post post = postRepository.findById(postId).orElseThrow(
+                () -> new IllegalArgumentException("게시글을 찾을 수 없습니다.")
+        );
+        System.out.println("a");
+        if (!post.validateUser(user)) {
+            throw new IllegalArgumentException("유저의 이름이 일치하지 않습니다.");
+        }
+        System.out.println("b");
+        long count = commentRepository.deleteByPost(post);
+        System.out.println(count);
         postRepository.delete(post);
     }
